Added happy pass tests

+++ src/main/java/co/inventorsoft/hoteloccupancy/dto/OccupancyRequestDto.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -17,5 +18,5 @@ public class OccupancyRequestDto {
 
     private Integer economyRooms;
 
-    private List<Double> potentialGuests;
+    private List<BigDecimal> potentialGuests;
 }
